+++ meinberlin/apps/geodata/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+
+
+class AddressSearchSerializer(serializers.Serializer):
+    search = serializers.CharField(required=False, allow_blank=True)
+    limit = serializers.IntegerField(default=100, min_value=1, max_value=1000)
+    offset = serializers.IntegerField(default=0, min_value=0)