+++ meinberlin/apps/plans/serializers.py
@@ -26,7 +26,7 @@ class PlanSerializer(PointSerializerMixin, serializers.ModelSerializer, CommonFi
     status = serializers.SerializerMethodField()
 
     def get_geojson_properties(self):
-        return {"strname": "street_name", "hsnr": "house_number", "plz": "zip_code"}
+        return {"str_name": "street_name", "hnr": "house_number", "plz": "zip_code"}
 
     class Meta:
         model = Plan