+++ meinberlin/apps/projects/forms.py
@@ -117,4 +117,9 @@ class PointForm(PointFormMixin, ProjectDashboardForm):
         }
 
     def get_geojson_properties(self):
-        return {"strname": "street_name", "hsnr": "house_number", "plz": "zip_code"}
+        return {
+            "str_name": "street_name",
+            "hnr": "house_number",
+            "plz": "zip_code",
+            "bez_name": "administrative_district.name",
+        }